+++ app/assets/js/models/contextualDates/index.tsx
@@ -1,14 +1,14 @@
 import { ContextualDate } from "@/api";
 import * as Time from "@/utils/time";
-import { DatePicker } from "turboui";
+import { DateField } from "turboui";
 
 /**
- * Takes a ContextualDate object and returns a DatePicker.ContextualDate object that can be used in the UI.
+ * Takes a ContextualDate object and returns a DateField.ContextualDate object that can be used in the UI.
  *
  * @param {ContextualDate | null | undefined} obj The ContextualDate object to parse.
- * @returns {DatePicker.ContextualDate | undefined} The DatePicker.ContextualDate object that was parsed, or undefined if the input was null or undefined.
+ * @returns {DateField.ContextualDate | undefined} The DateField.ContextualDate object that was parsed, or undefined if the input was null or undefined.
  */
-export function parseContextualDate(obj: ContextualDate | undefined | null): DatePicker.ContextualDate | null {
+export function parseContextualDate(obj: ContextualDate | undefined | null): DateField.ContextualDate | null {
   if (!obj) return null;
 
   return {
@@ -19,12 +19,12 @@ export function parseContextualDate(obj: ContextualDate | undefined | null): Dat
 }
 
 /**
- * Takes a DatePicker.ContextualDate object and returns a ContextualDate object that can be sent to the server.
+ * Takes a DateField.ContextualDate object and returns a ContextualDate object that can be sent to the server.
  *
- * @param {DatePicker.ContextualDate | null | undefined} obj The DatePicker.ContextualDate object to serialize.
+ * @param {DateField.ContextualDate | null | undefined} obj The DateField.ContextualDate object to serialize.
  * @returns {ContextualDate | null} The ContextualDate object that was serialized, or null if the input was null or undefined.
  */
-export function serializeContextualDate(obj: DatePicker.ContextualDate | undefined | null): ContextualDate | null {
+export function serializeContextualDate(obj: DateField.ContextualDate | undefined | null): ContextualDate | null {
   if (!obj) return null;
 
   return {
