+++ turboui/src/DateField/components/YearSelector.tsx
@@ -1,11 +1,11 @@
 import React, { useRef, useLayoutEffect } from "react";
 import { OptionButton } from "./OptionButton";
-import { DatePicker } from "../index";
+import { DateField } from "../index";
 import { getYearDate } from "../utils";
 
 interface Props {
-  selectedDate: DatePicker.ContextualDate | null;
-  setSelectedDate: React.Dispatch<React.SetStateAction<DatePicker.ContextualDate>>;
+  selectedDate: DateField.ContextualDate | null;
+  setSelectedDate: React.Dispatch<React.SetStateAction<DateField.ContextualDate>>;
   years: number[];
   useStartOfPeriod?: boolean;
 }
@@ -63,7 +63,7 @@ export function YearSelector({ selectedDate, setSelectedDate, years, useStartOfP
   );
 }
 
-function isSelectedYear(year: number, selectedDate: DatePicker.ContextualDate | null): boolean {
+function isSelectedYear(year: number, selectedDate: DateField.ContextualDate | null): boolean {
   if (!selectedDate?.date) return false;
 
   return selectedDate.dateType === "year" && selectedDate.date.getFullYear() === year;
