+++ turboui/src/DateField/index.tsx
@@ -1,134 +1,296 @@
-import * as Popover from "@radix-ui/react-popover";
-import { IconCalendarEvent, IconCalendarPlus, IconX } from "../icons";
 import React, { useState } from "react";
-import ReactDatePicker from "react-datepicker";
-import { SecondaryButton } from "../Button";
-import { createTestId } from "../TestableElement";
+import * as Popover from "@radix-ui/react-popover";
+import { IconCalendarEvent, IconX } from "../icons";
+import { InlineCalendar } from "./components/InlineCalendar";
+import DateTypeSelector from "./components/DateTypeSelector";
+import { MonthSelector } from "./components/MonthSelector";
+import { QuarterSelector } from "./components/QuarterSelector";
+import { YearSelector } from "./components/YearSelector";
+import ActionButtons from "./components/ActionButtons";
 import classNames from "../utils/classnames";
-import { isCurrentYear, isOverdue } from "../utils/time";
+import { isOverdue } from "../utils/time";
+import { createTestId, TestableElement } from "../TestableElement";
+
+const DATE_TYPES = [
+  { value: "day" as const, label: "Day" },
+  { value: "month" as const, label: "Month" },
+  { value: "quarter" as const, label: "Quarter" },
+  { value: "year" as const, label: "Year" },
+];
 
 export namespace DateField {
-  export interface Props {
-    date: Date | null;
-    setDate?: (date: Date | null) => void;
-
+  export interface Props extends TestableElement {
+    onDateSelect?: (selectedDate: ContextualDate | null) => void;
+    onCancel?: () => void;
+    date?: ContextualDate | null;
+    minYear?: number;
+    maxYear?: number;
     placeholder?: string;
     readonly?: boolean;
-    className?: string;
-    iconSize?: number;
-    textSize?: string;
-
     showOverdueWarning?: boolean;
-    showEmptyStateAsButton?: boolean;
-    showIcon?: boolean;
-    emptyStateText?: string;
-    emptyStateReadonlyText?: string;
-
     variant?: "inline" | "form-field";
-    testId?: string;
+    hideCalendarIcon?: boolean;
+    useStartOfPeriod?: boolean;
+  }
+
+  export type DateType = "day" | "month" | "quarter" | "year";
+
+  export interface ContextualDate {
+    date: Date;
+    dateType: DateType;
+    value: string;
+  }
+
+  export interface DateTypeOption {
+    value: DateType;
+    label: string;
+  }
+
+  export interface MonthOption {
+    value: string;
+    label: string;
+    name: string;
+  }
+
+  export interface PeriodOption {
+    value: string;
+    label: string;
   }
 }
 
 export function DateField({
+  onDateSelect,
+  onCancel,
   date,
-  setDate = () => {},
+  minYear = 2020,
+  maxYear = 2030,
+  placeholder = "Date",
   readonly = false,
-  iconSize = 18,
-  textSize = "text-sm",
+  showOverdueWarning = false,
   variant = "inline",
-  className = "",
-  showOverdueWarning = true,
-  showEmptyStateAsButton = false,
-  showIcon = true,
-  emptyStateText = "Set date",
-  emptyStateReadonlyText = "No date set",
+  hideCalendarIcon = false,
+  useStartOfPeriod = false,
   testId = "date-field",
 }: DateField.Props) {
-  const [isOpen, setIsOpen] = useState(false);
+  const [open, setOpen] = useState(false);
+  const [selectedDate, setSelectedDate] = useState<DateField.ContextualDate | null>(date || null);
+  const [previousSelectedDate, setPreviousSelectedDate] = useState<DateField.ContextualDate | null>(date || null);
+  const [dateType, setDateType] = useState<DateField.DateType>(date?.dateType || "day");
 
-  const handleChange = (newDate: Date | null) => {
-    if (date?.getTime() !== newDate?.getTime()) {
-      setDate(newDate);
+  const yearOptions = Array.from({ length: maxYear - minYear + 1 }, (_, i) => minYear + i);
+
+  const handleOpenChange = (isOpen: boolean) => {
+    setOpen(isOpen);
+
+    if (isOpen && selectedDate) {
+      setPreviousSelectedDate(selectedDate);
     }
-    setIsOpen(false); // Always close popover after date selection
   };
 
-  const clearDate = () => {
-    handleChange(null);
-    setIsOpen(false);
+  const handleTriggerClick = () => {
+    if (readonly) return;
+
+    setOpen(!open);
+
+    if (!open) {
+      setPreviousSelectedDate(selectedDate);
+    }
   };
 
-  useExportedTestHelper(testId, { setDate: handleChange });
+  const handleCancel = () => {
+    setSelectedDate(previousSelectedDate);
+    setOpen(false);
+    onCancel?.();
+  };
 
-  const triggerClassName = classNames(
-    "inline-block focus:outline-none focus:ring-2 focus:ring-primary-base",
-    {
-      "hover:bg-surface-dimmed rounded-lg px-1.5 py-1 -mx-1.5 -my-1": variant === "inline" && !readonly,
-      "border border-surface-outline rounded-lg w-full hover:bg-surface-dimmed px-2 py-1.5": variant === "form-field",
-    },
-    className,
-  );
+  const handleDateSelect = () => {
+    setOpen(false);
+
+    if (selectedDate) {
+      onDateSelect?.(selectedDate);
+    } else {
+      onDateSelect?.(null);
+    }
+  };
+
+  const handleClearDate = () => {
+    setSelectedDate(null);
+    setPreviousSelectedDate(null);
+    onDateSelect?.(null);
+    setDateType("day");
+    setOpen(false);
+  };
 
   return (
-    <Popover.Root open={isOpen} onOpenChange={setIsOpen}>
-      <Popover.Trigger 
-        className={triggerClassName} 
-        disabled={readonly} 
-        data-test-id={testId}
-        onKeyDown={(e) => {
-          if (e.key === "Enter" && !readonly) {
-            e.preventDefault();
-            setIsOpen(true);
-          }
-        }}
-      >
-        <DateDisplay
-          date={date}
-          className={className}
-          readonly={readonly}
-          showOverdueWarning={showOverdueWarning}
-          showEmptyStateAsButton={showEmptyStateAsButton}
-          showIcon={showIcon}
-          emptyStateText={emptyStateText}
-          emptyStateReadonlyText={emptyStateReadonlyText}
-          iconSize={iconSize}
-          textSize={textSize}
-          variant={variant}
-        />
+    <Popover.Root open={open} onOpenChange={handleOpenChange}>
+      <Popover.Trigger asChild disabled={readonly}>
+        <div>
+          <DatePickerTrigger
+            selectedDate={selectedDate}
+            onClick={handleTriggerClick}
+            label={placeholder}
+            readonly={readonly}
+            showOverdueWarning={showOverdueWarning}
+            variant={variant}
+            hideCalendarIcon={hideCalendarIcon}
+            testId={testId}
+          />
+        </div>
       </Popover.Trigger>
 
       <Popover.Portal>
-        <Popover.Content
-          className="bg-surface-base shadow-lg border border-surface-outline rounded-md z-[60]"
-          sideOffset={5}
-          onKeyDown={(e) => {
-            if (e.key === "Escape") {
-              e.preventDefault();
-              e.stopPropagation();
-              setIsOpen(false);
-            }
-          }}
-        >
-          <div className="flex justify-between items-center border-b border-stroken-base p-2 pb-1.5">
-            <div className="text-sm font-medium">Select date</div>
-            {date && <ClearButton clearDate={clearDate} testId={testId} />}
-          </div>
-
-          <div className="p-2 pt-1" data-test-id={createTestId(testId, "datepicker")}>
-            <ReactDatePicker selected={date || undefined} onChange={handleChange} inline />
-          </div>
-
-          <Popover.Arrow />
+        <Popover.Content className="z-50 animate-fadeIn" sideOffset={5} align="start">
+          <DatePickerContent
+            dateType={dateType}
+            setDateType={setDateType}
+            selectedDate={selectedDate}
+            setSelectedDate={setSelectedDate}
+            onDateSelect={handleDateSelect}
+            onCancel={handleCancel}
+            onClearDate={handleClearDate}
+            yearOptions={yearOptions}
+            testId={testId}
+            useStartOfPeriod={useStartOfPeriod}
+          />
         </Popover.Content>
       </Popover.Portal>
     </Popover.Root>
   );
 }
 
-function ClearButton({ clearDate, testId }: { clearDate: () => void; testId: string }) {
+interface DatePickerTriggerProps extends TestableElement {
+  selectedDate: DateField.ContextualDate | null;
+  label: string;
+  onClick: () => void;
+  readonly: boolean;
+  showOverdueWarning: boolean;
+  variant: "inline" | "form-field";
+  hideCalendarIcon: boolean;
+}
+
+function DatePickerTrigger({
+  selectedDate,
+  label,
+  onClick,
+  readonly,
+  showOverdueWarning,
+  variant,
+  hideCalendarIcon,
+  testId,
+}: DatePickerTriggerProps) {
+  const triggerClassName = classNames(
+    "inline-block focus:outline-none focus:ring-2 focus:ring-primary-base",
+    variant === "inline"
+      ? "rounded-lg px-1.5 py-1 -mx-1.5 -my-1"
+      : "border border-surface-outline rounded-lg w-full px-2 py-1.5",
+    !readonly ? "hover:bg-surface-dimmed" : "",
+  );
+
+  const handleClick = (e: React.MouseEvent) => {
+    e.preventDefault();
+    onClick();
+  };
+
+  let displayText = selectedDate?.value || label;
+  const isDateOverdue = selectedDate?.date && isOverdue(selectedDate.date);
+
+  const elemClass = classNames(
+    {
+      "flex items-center": true,
+      "gap-1.5": true,
+      "text-content-error": isDateOverdue && showOverdueWarning,
+      "text-content-dimmed": !selectedDate,
+    },
+    "text-sm",
+  );
+
   return (
     <button
-      onClick={clearDate}
+      type="button"
+      className={triggerClassName}
+      onClick={handleClick}
+      disabled={readonly}
+      aria-readonly={readonly}
+      data-testid={testId}
+    >
+      <span className={elemClass}>
+        {!hideCalendarIcon && (
+          <IconCalendarEvent
+            size={16}
+            className={isDateOverdue && showOverdueWarning ? "text-content-error -mt-[1px]" : "-mt-[1px]"}
+          />
+        )}
+        <span className="truncate">{displayText}</span>
+      </span>
+    </button>
+  );
+}
+
+interface DatePickerContentProps {
+  dateType: DateField.DateType;
+  setDateType: React.Dispatch<React.SetStateAction<DateField.DateType>>;
+  selectedDate: DateField.ContextualDate | null;
+  setSelectedDate: React.Dispatch<React.SetStateAction<DateField.ContextualDate | null>>;
+  onDateSelect?: (selectedDate: DateField.ContextualDate | null) => void;
+  onCancel?: () => void;
+  onClearDate?: () => void;
+  yearOptions: number[];
+  testId: string;
+  useStartOfPeriod?: boolean;
+}
+
+function DatePickerContent(props: DatePickerContentProps) {
+  const {
+    dateType,
+    setDateType,
+    selectedDate,
+    setSelectedDate,
+    onDateSelect,
+    onCancel,
+    onClearDate,
+    yearOptions,
+    testId,
+  } = props;
+
+  return (
+    <div className="max-w-md min-w-[300px] p-6 bg-surface-base rounded-lg shadow-lg">
+      <div className="flex justify-between items-center mb-4">
+        <h2 className="text-lg font-semibold flex items-center gap-2">
+          <IconCalendarEvent size={19} />
+          Set Date
+        </h2>
+        {selectedDate && <ClearButton onClear={onClearDate} testId={testId} />}
+      </div>
+
+      <DateTypeSelector dateType={dateType} dateTypes={DATE_TYPES} setDateType={setDateType} />
+
+      {dateType === "day" && (
+        <div className="mb-3">
+          <InlineCalendar selectedDate={selectedDate} setSelectedDate={setSelectedDate} />
+        </div>
+      )}
+
+      {dateType === "month" && (
+        <MonthSelector selectedDate={selectedDate} setSelectedDate={setSelectedDate} visibleYears={yearOptions} useStartOfPeriod={props.useStartOfPeriod} />
+      )}
+
+      {dateType === "quarter" && (
+        <QuarterSelector selectedDate={selectedDate} setSelectedDate={setSelectedDate} visibleYears={yearOptions} useStartOfPeriod={props.useStartOfPeriod} />
+      )}
+
+      {dateType === "year" && (
+        <YearSelector selectedDate={selectedDate} setSelectedDate={setSelectedDate} years={yearOptions} useStartOfPeriod={props.useStartOfPeriod} />
+      )}
+
+      <ActionButtons selectedDate={selectedDate} onCancel={onCancel} onSetDeadline={onDateSelect} />
+    </div>
+  );
+}
+
+function ClearButton({ onClear, testId }: { onClear?: () => void; testId: string }) {
+  return (
+    <button
+      onClick={() => onClear?.()}
       className="flex items-center text-xs text-content-subtle px-2 py-1 rounded hover:bg-surface-dimmed"
       data-test-id={createTestId(testId, "clear")}
     >
@@ -136,124 +298,4 @@ function ClearButton({ clearDate, testId }: { clearDate: () => void; testId: str
       Clear
     </button>
   );
-}
-
-interface DateDisplayProps {
-  date: Date | null | undefined;
-  className: string;
-  readonly: boolean;
-
-  iconSize: number;
-  textSize: string;
-
-  showOverdueWarning: boolean;
-  showEmptyStateAsButton: boolean;
-  showIcon: boolean;
-
-  emptyStateText: string;
-  emptyStateReadonlyText: string;
-
-  variant?: "inline" | "form-field";
-}
-
-function DateDisplay(props: DateDisplayProps) {
-  if (!props.date && props.showEmptyStateAsButton && props.readonly) {
-    return <EmptyStateButton emptyStateText={props.emptyStateText} readonly={props.readonly} variant={props.variant} />;
-  }
-
-  const iconSize = props.iconSize;
-  const textSize = props.textSize;
-  const isDateOverdue = props.date && isOverdue(props.date);
-  const variant = props.variant || "inline";
-
-  const elemClass = classNames(
-    {
-      "flex items-center": true,
-      "gap-1.5": props.showIcon,
-      "text-content-error": isDateOverdue && props.showOverdueWarning,
-      "text-content-dimmed": !props.date,
-      "w-full": variant === "form-field",
-    },
-    textSize,
-    props.className,
-  );
-
-  let text = "";
-  if (props.date) {
-    text = formatDate(props.date);
-  } else if (props.readonly) {
-    text = props.emptyStateReadonlyText;
-  } else {
-    text = props.emptyStateText;
-  }
-
-  return (
-    <span className={elemClass}>
-      {props.showIcon && <IconCalendarEvent size={iconSize} className="-mt-[1px]" />}
-      <span>{text}</span>
-    </span>
-  );
-}
-
-function EmptyStateButton({
-  readonly,
-  emptyStateText,
-  variant,
-}: {
-  readonly: boolean;
-  emptyStateText: string;
-  variant?: "inline" | "form-field";
-}) {
-  if (readonly) {
-    return null;
-  } else {
-    // If it's a form-field variant, we'll add padding to make it look like a form field
-    const containerClass = classNames({
-      "text-content-subtle": true,
-      "p-1.5": variant === "form-field",
-      "w-full": variant === "form-field",
-    });
-
-    // For form-field variant, wrap the button in a container that takes full width
-    return (
-      <div className={containerClass}>
-        <SecondaryButton size="xs" icon={IconCalendarPlus}>
-          {emptyStateText}
-        </SecondaryButton>
-      </div>
-    );
-  }
-}
-
-const formatDate = (date: Date) => {
-  if (isCurrentYear(date)) {
-    return date.toLocaleDateString("en-US", { month: "short", day: "numeric" });
-  } else {
-    return date.toLocaleDateString("en-US", { month: "short", day: "numeric", year: "numeric" });
-  }
-};
-
-//
-// This function allows tests to set the date programmatically
-// It uses useImperativeHandle to expose a method that can be called from tests
-//
-// To use this in tests, you would do something this:
-//
-// window.__tests.components["<testId>"].setDateFromTests(new Date("2023-10-01"));
-//
-
-function useExportedTestHelper(testId: string, api: any) {
-  React.useEffect(() => {
-    if (typeof window !== "undefined") {
-      window.__tests = window.__tests || {};
-      window.__tests.components = window.__tests.components || {};
-      window.__tests.components[testId] = api;
-    }
-
-    return () => {
-      if (window.__tests && window.__tests.components && window.__tests.components[testId]) {
-        delete window.__tests.components[testId];
-      }
-    };
-  }, [testId]);
 }
